Refactor tensor access and type casting in postprocessing modules
- Updated tensor access in RF-DETR, YOLO, and RT-DETR postprocessors to use static_cast for index conversions, ensuring type safety and clarity.
- Enhanced readability by consistently applying static_cast for size_t conversions across various tensor operations.
- Improved handling of class IDs and confidence scores by explicitly casting to float where necessary.
- Adjusted logic in video classification postprocessor to use size_t for loop indices, enhancing type safety.
- Minor adjustments in optical flow postprocessor for consistent type casting in calculations and color wheel generation.

+++ src/classification/torchvision_postprocessor.cpp
@@ -40,7 +40,7 @@ std::vector<Classification> TorchvisionPostprocessor::postprocess(const std::vec
 
     std::vector<Classification> classifications;
     int limit = std::min(top_k_, static_cast<int>(indexed_scores.size()));
-    for (int i = 0; i < limit; ++i) {
+    for (size_t i = 0; i < static_cast<size_t>(limit); ++i) {
         Classification cls;
         cls.class_id = static_cast<float>(indexed_scores[i].first);
         cls.class_confidence = indexed_scores[i].second;
