+++ src/object_detection/yolo_postprocessor.cpp
@@ -14,14 +14,14 @@ YoloPostprocessor::YoloPostprocessor(ObjectDetectionTask::ModelType model_type, 
 cv::Rect YoloPostprocessor::scaleToOriginal(float cx, float cy, float w, float h, const cv::Size& frame_size) const {
     // Apply letterbox inverse transformation
     // This converts coordinates from letterboxed model space to original frame space
-    float r_w = static_cast<float>(input_size_.width) / frame_size.width;
-    float r_h = static_cast<float>(input_size_.height) / frame_size.height;
+    float r_w = static_cast<float>(input_size_.width) / static_cast<float>(frame_size.width);
+    float r_h = static_cast<float>(input_size_.height) / static_cast<float>(frame_size.height);
 
     int x, y, width, height;
 
     if (r_h > r_w) {
         // Width is the limiting factor - padding is on top/bottom
-        float pad_h = (input_size_.height - r_w * frame_size.height) / 2.0f;
+        float pad_h = (static_cast<float>(input_size_.height) - r_w * static_cast<float>(frame_size.height)) / 2.0f;
         float x_min = cx - w / 2.0f;
         float x_max = cx + w / 2.0f;
         float y_min = cy - h / 2.0f - pad_h;
@@ -33,7 +33,7 @@ cv::Rect YoloPostprocessor::scaleToOriginal(float cx, float cy, float w, float h
         height = static_cast<int>((y_max - y_min) / r_w);
     } else {
         // Height is the limiting factor - padding is on left/right
-        float pad_w = (input_size_.width - r_h * frame_size.width) / 2.0f;
+        float pad_w = (static_cast<float>(input_size_.width) - r_h * static_cast<float>(frame_size.width)) / 2.0f;
         float x_min = cx - w / 2.0f - pad_w;
         float x_max = cx + w / 2.0f - pad_w;
         float y_min = cy - h / 2.0f;
@@ -109,8 +109,8 @@ std::vector<Detection> YoloPostprocessor::postprocessYoloStandard(const std::vec
         return {};
 
     // int batch = shape[0]; // Unused
-    int dim1 = shape[1];
-    int dim2 = shape[2];
+    int dim1 = static_cast<int>(shape[1]);
+    int dim2 = static_cast<int>(shape[2]);
 
     // Detect format by shape:
     // YOLOv5/v6/v7: [batch, anchors, 4+1+classes] where dim2 < dim1, has objectness at index 4
@@ -138,7 +138,7 @@ std::vector<Detection> YoloPostprocessor::postprocessYoloStandard(const std::vec
         // For YOLOv5/v6/v7, check objectness score first (index 4)
         float objectness = 1.0f;
         if (has_objectness) {
-            objectness = getTensorFloat(output[i * channels + 4]);
+            objectness = getTensorFloat(output[static_cast<size_t>(i * channels + 4)]);
             if (objectness < confidence_threshold_)
                 continue;
         }
@@ -151,10 +151,10 @@ std::vector<Detection> YoloPostprocessor::postprocessYoloStandard(const std::vec
             float score;
             if (has_objectness) {
                 // YOLOv5/v6/v7: [batch, anchors, channels] -> data[i * channels + (c + 5)]
-                score = getTensorFloat(output[i * channels + (c + class_offset)]);
+                score = getTensorFloat(output[static_cast<size_t>(i * channels + (c + class_offset))]);
             } else {
                 // YOLOv8+: [batch, channels, anchors] -> data[(c + 4) * anchors + i]
-                score = getTensorFloat(output[(c + class_offset) * anchors + i]);
+                score = getTensorFloat(output[static_cast<size_t>((c + class_offset) * anchors + i)]);
             }
 
             if (score > max_class_score) {
@@ -173,20 +173,20 @@ std::vector<Detection> YoloPostprocessor::postprocessYoloStandard(const std::vec
         float cx, cy, w, h;
         if (has_objectness) {
             // YOLOv5/v6/v7: [batch, anchors, channels]
-            cx = getTensorFloat(output[i * channels + 0]);
-            cy = getTensorFloat(output[i * channels + 1]);
-            w = getTensorFloat(output[i * channels + 2]);
-            h = getTensorFloat(output[i * channels + 3]);
+            cx = getTensorFloat(output[static_cast<size_t>(i * channels + 0)]);
+            cy = getTensorFloat(output[static_cast<size_t>(i * channels + 1)]);
+            w = getTensorFloat(output[static_cast<size_t>(i * channels + 2)]);
+            h = getTensorFloat(output[static_cast<size_t>(i * channels + 3)]);
         } else {
             // YOLOv8+: [batch, channels, anchors]
-            cx = getTensorFloat(output[0 * anchors + i]);
-            cy = getTensorFloat(output[1 * anchors + i]);
-            w = getTensorFloat(output[2 * anchors + i]);
-            h = getTensorFloat(output[3 * anchors + i]);
+            cx = getTensorFloat(output[static_cast<size_t>(0 * anchors + i)]);
+            cy = getTensorFloat(output[static_cast<size_t>(1 * anchors + i)]);
+            w = getTensorFloat(output[static_cast<size_t>(2 * anchors + i)]);
+            h = getTensorFloat(output[static_cast<size_t>(3 * anchors + i)]);
         }
 
         Detection det;
-        det.class_id = class_id;
+        det.class_id = static_cast<float>(class_id);
         det.class_confidence = final_score;
         det.bbox = scaleToOriginal(cx, cy, w, h, frame_size);
         detections.push_back(det);
@@ -222,14 +222,14 @@ std::vector<Detection> YoloPostprocessor::postprocessYoloV4(const std::vector<Te
                 float w = getTensorFloat(output[2]);
                 float h = getTensorFloat(output[3]);
 
-                int label = maxSPtr - (output + 5);
+                int label = static_cast<int>(maxSPtr - (output + 5));
 
                 // YOLOv4 outputs normalized coordinates [0,1] relative to the model
                 // input Convert to model input pixel coordinates
-                float cx_model = cx * input_size_.width;
-                float cy_model = cy * input_size_.height;
-                float w_model = w * input_size_.width;
-                float h_model = h * input_size_.height;
+                float cx_model = cx * static_cast<float>(input_size_.width);
+                float cy_model = cy * static_cast<float>(input_size_.height);
+                float w_model = w * static_cast<float>(input_size_.width);
+                float h_model = h * static_cast<float>(input_size_.height);
 
                 // Apply letterbox inverse transformation using model-space coordinates
                 cv::Rect bbox = scaleToOriginal(cx_model, cy_model, w_model, h_model, frame_size);
@@ -238,7 +238,7 @@ std::vector<Detection> YoloPostprocessor::postprocessYoloV4(const std::vector<Te
                 Detection detection;
                 detection.bbox = bbox;
                 detection.class_confidence = score;
-                detection.class_id = label;
+                detection.class_id = static_cast<float>(label);
 
                 // Add the detection to the vector
                 detections.push_back(detection);
@@ -251,7 +251,7 @@ std::vector<Detection> YoloPostprocessor::postprocessYoloV4(const std::vector<Te
     std::map<int, std::vector<size_t>> class2indices;
     for (size_t i = 0; i < detections.size(); i++) {
         if (detections[i].class_confidence >= confidence_threshold_) {
-            class2indices[detections[i].class_id].push_back(i);
+            class2indices[static_cast<int>(detections[i].class_id)].push_back(i);
         }
     }
 
@@ -266,7 +266,7 @@ std::vector<Detection> YoloPostprocessor::postprocessYoloV4(const std::vector<Te
         std::vector<int> nmsIndices;
         cv::dnn::NMSBoxes(localBoxes, localConfidences, confidence_threshold_, nms_threshold_, nmsIndices);
         for (size_t i = 0; i < nmsIndices.size(); i++) {
-            filtered_detections.push_back(detections[classIndices[nmsIndices[i]]]);
+            filtered_detections.push_back(detections[classIndices[static_cast<size_t>(nmsIndices[i])]]);
         }
     }
 
@@ -281,33 +281,33 @@ std::vector<Detection> YoloPostprocessor::postprocessYoloNmsFree(const Tensor& o
     if (output.shape.size() < 3 || output.shape[2] < 6)
         return {};
 
-    int num_dets = output.shape[1];
-    int dims = output.shape[2];
+    int num_dets = static_cast<int>(output.shape[1]);
+    int dims = static_cast<int>(output.shape[2]);
 
     // Calculate scale ratios for letterbox inverse
-    float r_w = static_cast<float>(input_size_.width) / frame_size.width;
-    float r_h = static_cast<float>(input_size_.height) / frame_size.height;
+    float r_w = static_cast<float>(input_size_.width) / static_cast<float>(frame_size.width);
+    float r_h = static_cast<float>(input_size_.height) / static_cast<float>(frame_size.height);
 
     for (int i = 0; i < num_dets; ++i) {
-        float score = getTensorFloat(output.data[i * dims + 4]);
+        float score = getTensorFloat(output.data[static_cast<size_t>(i * dims + 4)]);
         if (score < confidence_threshold_)
             continue;
 
-        float x1 = getTensorFloat(output.data[i * dims + 0]);
-        float y1 = getTensorFloat(output.data[i * dims + 1]);
-        float x2 = getTensorFloat(output.data[i * dims + 2]);
-        float y2 = getTensorFloat(output.data[i * dims + 3]);
-        int class_id = static_cast<int>(getTensorFloat(output.data[i * dims + 5]));
+        float x1 = getTensorFloat(output.data[static_cast<size_t>(i * dims + 0)]);
+        float y1 = getTensorFloat(output.data[static_cast<size_t>(i * dims + 1)]);
+        float x2 = getTensorFloat(output.data[static_cast<size_t>(i * dims + 2)]);
+        float y2 = getTensorFloat(output.data[static_cast<size_t>(i * dims + 3)]);
+        int class_id = static_cast<int>(getTensorFloat(output.data[static_cast<size_t>(i * dims + 5)]));
 
         // Apply letterbox inverse transformation
         if (r_h > r_w) {
-            float pad_h = (input_size_.height - r_w * frame_size.height) / 2.0f;
+            float pad_h = (static_cast<float>(input_size_.height) - r_w * static_cast<float>(frame_size.height)) / 2.0f;
             y1 = (y1 - pad_h) / r_w;
             y2 = (y2 - pad_h) / r_w;
             x1 = x1 / r_w;
             x2 = x2 / r_w;
         } else {
-            float pad_w = (input_size_.width - r_h * frame_size.width) / 2.0f;
+            float pad_w = (static_cast<float>(input_size_.width) - r_h * static_cast<float>(frame_size.width)) / 2.0f;
             x1 = (x1 - pad_w) / r_h;
             x2 = (x2 - pad_w) / r_h;
             y1 = y1 / r_h;
@@ -315,7 +315,7 @@ std::vector<Detection> YoloPostprocessor::postprocessYoloNmsFree(const Tensor& o
         }
 
         Detection det;
-        det.class_id = class_id;
+        det.class_id = static_cast<float>(class_id);
         det.class_confidence = score;
         det.bbox =
             cv::Rect(static_cast<int>(x1), static_cast<int>(y1), static_cast<int>(x2 - x1), static_cast<int>(y2 - y1));
@@ -336,12 +336,12 @@ std::vector<Detection> YoloPostprocessor::postprocessYoloNAS(const Tensor& boxes
     if (boxes.shape.size() < 3 || scores.shape.size() < 3)
         return {};
 
-    int num_dets = boxes.shape[1];
-    int num_classes = scores.shape[2];
+    int num_dets = static_cast<int>(boxes.shape[1]);
+    int num_classes = static_cast<int>(scores.shape[2]);
 
     // Calculate scale ratios for letterbox inverse
-    float r_w = static_cast<float>(input_size_.width) / frame_size.width;
-    float r_h = static_cast<float>(input_size_.height) / frame_size.height;
+    float r_w = static_cast<float>(input_size_.width) / static_cast<float>(frame_size.width);
+    float r_h = static_cast<float>(input_size_.height) / static_cast<float>(frame_size.height);
 
     for (int i = 0; i < num_dets; ++i) {
         // Find max score for this detection
@@ -349,7 +349,7 @@ std::vector<Detection> YoloPostprocessor::postprocessYoloNAS(const Tensor& boxes
         int class_id = -1;
 
         for (int c = 0; c < num_classes; ++c) {
-            float score = getTensorFloat(scores.data[i * num_classes + c]);
+            float score = getTensorFloat(scores.data[static_cast<size_t>(i * num_classes + c)]);
             if (score > max_score) {
                 max_score = score;
                 class_id = c;
@@ -359,20 +359,20 @@ std::vector<Detection> YoloPostprocessor::postprocessYoloNAS(const Tensor& boxes
         if (max_score < confidence_threshold_)
             continue;
 
-        float x1 = getTensorFloat(boxes.data[i * 4 + 0]);
-        float y1 = getTensorFloat(boxes.data[i * 4 + 1]);
-        float x2 = getTensorFloat(boxes.data[i * 4 + 2]);
-        float y2 = getTensorFloat(boxes.data[i * 4 + 3]);
+        float x1 = getTensorFloat(boxes.data[static_cast<size_t>(i * 4 + 0)]);
+        float y1 = getTensorFloat(boxes.data[static_cast<size_t>(i * 4 + 1)]);
+        float x2 = getTensorFloat(boxes.data[static_cast<size_t>(i * 4 + 2)]);
+        float y2 = getTensorFloat(boxes.data[static_cast<size_t>(i * 4 + 3)]);
 
         // Apply letterbox inverse transformation
         if (r_h > r_w) {
-            float pad_h = (input_size_.height - r_w * frame_size.height) / 2.0f;
+            float pad_h = (static_cast<float>(input_size_.height) - r_w * static_cast<float>(frame_size.height)) / 2.0f;
             y1 = (y1 - pad_h) / r_w;
             y2 = (y2 - pad_h) / r_w;
             x1 = x1 / r_w;
             x2 = x2 / r_w;
         } else {
-            float pad_w = (input_size_.width - r_h * frame_size.width) / 2.0f;
+            float pad_w = (static_cast<float>(input_size_.width) - r_h * static_cast<float>(frame_size.width)) / 2.0f;
             x1 = (x1 - pad_w) / r_h;
             x2 = (x2 - pad_w) / r_h;
             y1 = y1 / r_h;
@@ -380,7 +380,7 @@ std::vector<Detection> YoloPostprocessor::postprocessYoloNAS(const Tensor& boxes
         }
 
         Detection det;
-        det.class_id = class_id;
+        det.class_id = static_cast<float>(class_id);
         det.class_confidence = max_score;
         det.bbox =
             cv::Rect(static_cast<int>(x1), static_cast<int>(y1), static_cast<int>(x2 - x1), static_cast<int>(y2 - y1));
@@ -406,8 +406,9 @@ void YoloPostprocessor::applyNMS(std::vector<Detection>& detections) {
                 continue;
 
             cv::Rect intersection = detections[i].bbox & detections[j].bbox;
-            float intersection_area = intersection.area();
-            float union_area = detections[i].bbox.area() + detections[j].bbox.area() - intersection_area;
+            float intersection_area = static_cast<float>(intersection.area());
+            float union_area = static_cast<float>(detections[i].bbox.area()) +
+                               static_cast<float>(detections[j].bbox.area()) - intersection_area;
 
             if (union_area > 0) {
                 float iou = intersection_area / union_area;
@@ -425,7 +426,7 @@ void YoloPostprocessor::applyNMS(std::vector<Detection>& detections) {
 
     detections.erase(std::remove_if(detections.begin(), detections.end(),
                                     [&](const Detection& det) {
-                                        size_t idx = &det - &detections[0];
+                                        size_t idx = static_cast<size_t>(&det - &detections[0]);
                                         return suppress[idx];
                                     }),
                      detections.end());
@@ -455,43 +456,43 @@ std::vector<Detection> YoloPostprocessor::postprocessYoloV7E2E(const Tensor& num
     int num_dets = static_cast<int>(getTensorFloat(num_dets_tensor.data[0]));
 
     // boxes.shape is [1, max_dets, 4]
-    int max_dets = boxes.shape.size() >= 2 ? boxes.shape[1] : 100;
+    int max_dets = static_cast<int>(boxes.shape.size() >= 2 ? boxes.shape[1] : 100);
     num_dets = std::min(num_dets, max_dets); // Clamp to max_dets
 
     // Calculate scale ratios for letterbox inverse
-    float r_w = static_cast<float>(input_size_.width) / frame_size.width;
-    float r_h = static_cast<float>(input_size_.height) / frame_size.height;
+    float r_w = static_cast<float>(input_size_.width) / static_cast<float>(frame_size.width);
+    float r_h = static_cast<float>(input_size_.height) / static_cast<float>(frame_size.height);
 
     for (int i = 0; i < num_dets; ++i) {
-        float score = getTensorFloat(scores.data[i]);
+        float score = getTensorFloat(scores.data[static_cast<size_t>(i)]);
         if (score < confidence_threshold_)
             continue;
 
         // Extract box coordinates in model space
-        float x1 = getTensorFloat(boxes.data[i * 4 + 0]);
-        float y1 = getTensorFloat(boxes.data[i * 4 + 1]);
-        float x2 = getTensorFloat(boxes.data[i * 4 + 2]);
-        float y2 = getTensorFloat(boxes.data[i * 4 + 3]);
+        float x1 = getTensorFloat(boxes.data[static_cast<size_t>(i * 4 + 0)]);
+        float y1 = getTensorFloat(boxes.data[static_cast<size_t>(i * 4 + 1)]);
+        float x2 = getTensorFloat(boxes.data[static_cast<size_t>(i * 4 + 2)]);
+        float y2 = getTensorFloat(boxes.data[static_cast<size_t>(i * 4 + 3)]);
 
         // Apply letterbox inverse transformation
         if (r_h > r_w) {
-            float pad_h = (input_size_.height - r_w * frame_size.height) / 2.0f;
+            float pad_h = (static_cast<float>(input_size_.height) - r_w * static_cast<float>(frame_size.height)) / 2.0f;
             y1 = (y1 - pad_h) / r_w;
             y2 = (y2 - pad_h) / r_w;
             x1 = x1 / r_w;
             x2 = x2 / r_w;
         } else {
-            float pad_w = (input_size_.width - r_h * frame_size.width) / 2.0f;
+            float pad_w = (static_cast<float>(input_size_.width) - r_h * static_cast<float>(frame_size.width)) / 2.0f;
             x1 = (x1 - pad_w) / r_h;
             x2 = (x2 - pad_w) / r_h;
             y1 = y1 / r_h;
             y2 = y2 / r_h;
         }
 
-        int class_id = static_cast<int>(getTensorFloat(classes.data[i]));
+        int class_id = static_cast<int>(getTensorFloat(classes.data[static_cast<size_t>(i)]));
 
         Detection det;
-        det.class_id = class_id;
+        det.class_id = static_cast<float>(class_id);
         det.class_confidence = score;
         det.bbox =
             cv::Rect(static_cast<int>(x1), static_cast<int>(y1), static_cast<int>(x2 - x1), static_cast<int>(y2 - y1));
